refactor: add ordertable package
