+++ src/main/java/kitchenpos/domain/Order.java
@@ -4,6 +4,7 @@ import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+import kitchenpos.ordertable.domain.OrderTable;
 
 @Table(name = "orders")
 @Entity
