+++ src/main/java/kitchenpos/domain/OrderRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.domain;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public interface OrderRepository {
     Order save(Order order);
