+++ src/main/java/kitchenpos/ordertable/domain/OrderTableRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.ordertable.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public interface OrderTableRepository {
     OrderTable save(OrderTable orderTable);
